requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==5.2.3
    # via
    #   -r requirements.in
    #   django-unfold
django-unfold==0.60.0
    # via -r requirements.in
sqlparse==0.5.3
    # via django
